# Replace the typed aggregate functions with simpler untyped ones