--- a/src/realm/dictionary.hpp
+++ b/src/realm/dictionary.hpp
@@ -239,7 +239,6 @@ public:
 
     // Overrides of ObjList:
     ObjKey get_key(size_t ndx) const final;
-    bool is_obj_valid(size_t ndx) const noexcept final;
     Obj get_object(size_t row_ndx) const final;
 
     // Overrides of CollectionBase, these simply forward to the underlying dictionary.