--- a/src/realm/object-store/results.cpp
+++ b/src/realm/object-store/results.cpp
@@ -403,7 +403,7 @@ util::Optional<Obj> Results::try_get(size_t row_ndx)
         case Mode::TableView:
             if (row_ndx >= m_table_view.size())
                 break;
-            return m_table_view.try_get_object(row_ndx);
+            return m_table_view.get_object(row_ndx);
     }
     return util::none;
 }