--- a/src/realm/object-store/results.hpp
+++ b/src/realm/object-store/results.hpp
@@ -399,7 +399,6 @@ private:
 
     template <typename AggregateFunction>
     util::Optional<Mixed> aggregate(ColKey column, const char* name, AggregateFunction&& func) REQUIRES(!m_mutex);
-    DataType prepare_for_aggregate(ColKey column, const char* name) REQUIRES(m_mutex);
 
     template <typename Fn>
     auto dispatch(Fn&&) const REQUIRES(!m_mutex);