--- a/src/realm/object-store/set.hpp
+++ b/src/realm/object-store/set.hpp
@@ -91,16 +91,6 @@ public:
 
     Set freeze(const std::shared_ptr<Realm>& realm) const;
 
-    // Get the min/max/average/sum of the given column
-    // All but sum() returns none when there are zero matching rows
-    // sum() returns 0,
-    // Throws UnsupportedColumnTypeException for sum/average on timestamp or non-numeric column
-    // Throws OutOfBoundsIndexException for an out-of-bounds column
-    util::Optional<Mixed> max(ColKey column = {}) const;
-    util::Optional<Mixed> min(ColKey column = {}) const;
-    util::Optional<Mixed> average(ColKey column = {}) const;
-    Mixed sum(ColKey column = {}) const;
-
     bool is_subset_of(const Collection& rhs) const;
     bool is_strict_subset_of(const Collection& rhs) const;
     bool is_superset_of(const Collection& rhs) const;