--- a/src/realm/table_ref.hpp
+++ b/src/realm/table_ref.hpp
@@ -29,8 +29,8 @@ class TableRef;
 
 class ConstTableRef {
 public:
-    ConstTableRef() noexcept {}
-    ConstTableRef(std::nullptr_t) noexcept {}
+    constexpr ConstTableRef() noexcept = default;
+    constexpr ConstTableRef(std::nullptr_t) noexcept {}
     ConstTableRef(const TableRef& other) noexcept;
 
     const Table* operator->() const;
@@ -91,14 +91,8 @@ protected:
 
 class TableRef : public ConstTableRef {
 public:
-    TableRef() noexcept
-        : ConstTableRef()
-    {
-    }
-    TableRef(std::nullptr_t) noexcept
-        : ConstTableRef()
-    {
-    }
+    constexpr TableRef() noexcept = default;
+    constexpr TableRef(std::nullptr_t) noexcept {}
 
     Table* operator->() const;
     Table& operator*() const;