… CHANGELOG.md
… src/realm/CMakeLists.txt
… src/realm/object-store/list.cpp
… src/realm/query.hpp
… src/realm/query_conditions_tpl.hpp
… src/realm/sync/subscriptions.cpp
… src/realm/table.cpp
… src/realm/table.hpp
… src/realm/table_tpl.hpp
… src/realm/table_view.cpp
… test/object-store/benchmarks/results.cpp
… test/object-store/dictionary.cpp
… test/object-store/primitive_list.cpp
… test/test_column_timestamp.cpp
… test/test_decimal128.cpp
… test/test_lang_bind_helper.cpp
… test/test_metrics.cpp
… test/test_query.cpp
… test/test_query2.cpp
… test/test_table.cpp
… test/test_table_view.cpp
… test/test_transform.cpp